refactor: move reminder claim orchestration to service layer
