+++ src/agendable/db/repos/reminders.py
@@ -9,7 +9,6 @@ from sqlalchemy.engine import CursorResult
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
-import agendable.db as db
 from agendable.db.models import (
     MeetingOccurrence,
     MeetingSeries,
@@ -71,26 +70,3 @@ class ReminderRepository(BaseRepository[Reminder]):
 
         claim_rowcount = cast(CursorResult[object], claim_result).rowcount
         return claim_rowcount == 1
-
-
-async def claim_reminder_attempt(
-    *,
-    reminder: Reminder,
-    now: datetime,
-    claim_lease_seconds: int,
-) -> bool:
-    async with db.SessionMaker() as claim_session:
-        reminder_repo = ReminderRepository(claim_session)
-        was_claimed = await reminder_repo.try_claim_attempt(
-            reminder_id=reminder.id,
-            expected_attempt_count=reminder.attempt_count,
-            now=now,
-            claim_lease_seconds=claim_lease_seconds,
-        )
-        if not was_claimed:
-            return False
-        await claim_session.commit()
-
-    reminder.attempt_count += 1
-    reminder.last_attempted_at = now
-    return True
