+++ src/agendable/services/__init__.py
@@ -25,6 +25,7 @@ from agendable.services.oidc_service import (
     resolve_oidc_link_resolution,
     resolve_oidc_login_resolution,
 )
+from agendable.services.reminder_claim_service import claim_reminder_attempt
 from agendable.services.reminder_delivery_service import run_due_reminders
 from agendable.services.series_service import create_series_with_occurrences
 
@@ -42,6 +43,7 @@ __all__ = [
     "MissingGoogleCalendarConnectionError",
     "OidcLinkResolution",
     "OidcLoginResolution",
+    "claim_reminder_attempt",
     "complete_occurrence_and_roll_forward",
     "create_series_with_occurrences",
     "provision_user_for_oidc",
